Added Crash-prevention with the sequence
If you created a sequence then teleport to another place, when you
play the sequence the game enters on a infinite loop because the
distance is too far. To solve this, we check if the current camera
position is far from the first point of the sequence, in the case that's
true we clean the sequence.