Fix jittering issues (#4).
The issue wasn't the scheduler but a math singularity. When the radius
(or distance between the two points) is too small, some jittering was
produced and the movement speed was quite slow since the distance is the
magnitude of the speed. We fix that by using a fixed distance so the
speed will be the same at every distance of these two points. The bug
was tracked down thanks to the help of Wackiger, thanks a lot!