better error handling, using tcpstream.
UdpSocket had the tendency to drop some packages and it made
the replay very choppy. Changing it for a TcpStream instantly
fixed the choppyness. There's still some delay it needs to be fixed but
the playback is definitely better.

Also, improved error handling, in the case the project did not connect
to the blender plugin it won't crash.